azure-resource-inventory/requirement.txt
azure.mgmt.subscription
azure.identity
azure.mgmt.compute